Issue 465 Move OpenShift related commands into the 'openshift' context
