+++ docs/minishift_openshift_service_list.md
@@ -1,20 +1,27 @@
-## minishift openshift
+## minishift openshift service list
 
-Interact with an Openshift Cluster
+Gets the URLs of the services in your local cluster.
 
 ### Synopsis
 
 
-Interact with an Openshift cluster, check available versions and modify master config resources
+Gets the URLs of the services in your local cluster.
 
 ```
-minishift openshift SUBCOMMAND [flags]
+minishift openshift service list [flags]
+```
+
+### Options
+
+```
+  -n, --namespace string   The namespace of the services.
 ```
 
 ### Options inherited from parent commands
 
 ```
       --alsologtostderr                  log to standard error as well as files
+      --format string                    The URL format of the service. (default "http://{{.IP}}:{{.Port}}")
       --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
       --log_dir string                   If non-empty, write log files in this directory (default "")
       --logtostderr                      log to standard error instead of files
@@ -25,9 +32,5 @@ minishift openshift SUBCOMMAND [flags]
 ```
 
 ### SEE ALSO
-* [minishift](minishift.md)	 - Minishift is a tool for application development in local OpenShift clusters.
-* [minishift openshift config](minishift_openshift_config.md)	 - Displays or patches OpenShift configuration.
-* [minishift openshift list-version](minishift_openshift_list-version.md)	 - Gets the list of OpenShift versions available for Minishift.
-* [minishift openshift restart](minishift_openshift_restart.md)	 - Restarts the OpenShift cluster.
 * [minishift openshift service](minishift_openshift_service.md)	 - Prints the URL for the specified service to the console.
